+++ src/api/consumer/lib/login.js
@@ -3,7 +3,9 @@ import loginToBitSrc from '../../../consumer/login/login';
 
 export default (async function loginAction(
   port: string,
-  noLaunchBrowser?: boolean
+  noLaunchBrowser?: boolean,
+  rcPath: string,
+  skipRegistryConfig: boolean
 ): Promise<{ isAlreadyLoggedIn?: boolean, username?: string }> {
-  return loginToBitSrc(port, noLaunchBrowser);
+  return loginToBitSrc(port, noLaunchBrowser, rcPath, skipRegistryConfig);
 });